Small refactor (#211)
* chore: add instructio on how to checkout the submodule

* refactor: remove level of indirection

Call `groth16::create_random_proof` directly instead of having a helper
method. This reduces one level of indirection and 10 lines of code.

* chore: unify cache story

There are cache keys that aren't actually used and an instanceo of
a manually written cache key, which is error prone. This commit cleans
things a bit up.