refactor: avoid runtime mutability checks in the store (#491)
Since #364, the store relies on Rc<Refcell<_>> to ensure unique access to the store.

This is a runtime check that would lead to a panic should some other piece of code retain a reference to the store in a call to `get_expr`.
This PR eliminates the need for this unnecessary construct, and re-established the use of the borrow checker to ensure static memory safety.