Environment optimization (#1035)
* letrec works differently

letrec will now add thunk expressions which adds themselves to the
environment before evaluating

* environment tag added

* `Num`s now take arbitrary raw pointers

This is so that we can delay hashing, much like `Ptr`s delay hashing

* added PushBinding/PopBinding operations

* eval uses optimized environment

* updated tests to new environment

* reverted `letrec` change

Now `letrec` works as before, but updated to the new, changed,
environment. The reason for reverting is that the iteration count
increased in a bunch of tests. For thunks to work properly, we will
need the memoset implementation of eval

* test iterations updated

* fixed repl and fib benchmark

* fixed zstore

* fixed demo, except for protocol.lurk

* added empty-env operation and fixed protocol.lurk

* fixed cli test

* slightly more efficient lookup