Update sha256 coprocessor discarded bits
Previously, the pallas curve had a capacity of 254 bits. The new bn256
curve has a capacity of 253 bits. This makes the pack_bits gadget used by
sha256 coprocessor drop the three most significant bits of the output
instead of just two.

This commit updates the coprocessor's evaluate_simple implementation to
set the correct number of bits to zero based on the field used to avoid a
panic caused by a mismatch between the synthesize and evaluate outputs.