Fix coprocessors returning lists in SuperNova
In NIVC, the Lurk circuit calls out to the coprocessor circuit, and then
on the next fold step calls back into the main Lurk circuit with the
result of the coprocessor. If the coprocessor returns a single value,
there is no issue, but if the coprocessor returns a list, then the lurk
circuit attempts to evaluate that expression instead of simply returning
it. To fix it, we wrap the coprocessor return expr with a thunk.