feat: formalize "car_cdr_simple"
We already have an implementation of `car_cdr` in LEM that doesn't
deal with strings in order to save constraints. We can call that
`car_cdr_simple` so we know what we're talking about just by looking
at it.

This diff materializes this concept in the `Store` and implements
a corresponding gadget that can be used more cheaply in coprocessors.
Furthemore, `chain_car_cdr` gains a `simple` flag so the caller can
use it to save constraints.