refactor: Refactor proof handling for ownership and borrowing (#1214)
* refactor: Refactor proof handling for ownership and borrowing

On a long enough timeline, we'll want to generate compressed proofs without dropping the associated `RecursiveProof`. This is for the purpose of re-starting the proof using the prior
`RecursiveProof`, which at the moment can only be obtained through cloning prior to compression.

The current approach witnesses that compression only requires a reference to the current proof, and returns a `Cow`: the only situation in which this would result in a clone is if:
- we create a `CompressedProof`,
- we call `compress()` on it,
- we further perform a modificaiton of that `CompressedProof` that requires an owned value..

In detail:
- Updated `compress` function across various files to take `&self` instead of `self`, avoiding the drop of the `RecursiveProof`;
- Adjusted dependencies in chain-server's `Cargo.toml` to use workspace versions, streamlining dependency management and increasing project consistency.

* chore: clippy

* chore: update pointer to arecibo